+++ admin/assets/manageTravels/controller/add.php
@@ -12,14 +12,21 @@ if (isset($_POST)) {
             $title = $_POST['title'];
             $description = $_POST['description'];
             $category = $_POST['category'];
+            if (isset($_POST['display-carousel']) && $_POST['display-carousel'] == 'on') {
+                $display_carousel = '1';
+            }
+            else {
+                $display_carousel = '0';
+            }
 
-            $add = $bdd->prepare('INSERT INTO travels (addDate, picture, title, description, category_id) VALUES (:addDate, :picture, :title, :description, :category_id)');
+            $add = $bdd->prepare('INSERT INTO travels (addDate, picture, title, description, display_carousel, category_id) VALUES (:addDate, :picture, :title, :description, :display_carousel, :category_id)');
 
             if ($add->execute(array(
                 'addDate' => $addDate,
                 'picture' => $picture,
                 'title' => $title,
                 'description' => $description,
+                'display_carousel' => $display_carousel,
                 'category_id' => $category,
             ))) {
                 $confirmMessage = 'The article '.$title.' is well added in the database. Thanks for your add ! :D';
@@ -39,6 +46,7 @@ if (isset($_POST)) {
     if (isset($_GET['manageTravels']) && $_GET['manageTravels'] == 'category') {
         if (isset($_POST['category-list']) && $_POST['category-list'] != '') {
             $category = $_POST['category-list'];
+            $thumbnail = $_POST['thumbnail'];
         
             $add = $bdd->prepare('INSERT INTO categories_travel (category) VALUES (:category)');