+++ admin/assets/manageTravels/controller/update.php
@@ -1,0 +1,4 @@
+<?php 
+
+require('../../functions/functions-admin.php');
+